requirements.txt
pycryptodome
eyed3
tqdm
psutil
requests